\name{TAPS_call}
\alias{TAPS_call}
\alias{TAPS.call}
\alias{Taps.call}
\alias{Taps_call}


\title{
Tumor Abberation Prediction Suite
}

\description{
TAPS_call calls copynumbers and displays total and minor copynumbers of a tumor sample.
}

\usage{
TAPS_call(directory=NULL,minseg=1,maxCn=12)
}

\arguments{
	\item{directory}{
	Default is getwd().
	}
	\item{minseg}{
	Default is 1.
	}
	\item{maxCn}{
	Default is 12.
	}
}

\details{
TAPS_call is the second step in TAPS analysis. It takes two things as input. \cr
1) Your interpretation of plots which you add to the SampleData.txt file. (see homepage for further instructions at
http://patchwork.r-forge.r-project.org/) \cr
2) The .Rdata files generated by TAPS_plot().

%ADD MORE DETAILS HERE LATER!
}

\references{
http://www.biomedcentral.com/content/pdf/gb-2011-12-10-r108.pdf
}

\author{
Markus Mayrhofer, \email{markus.mayrhofer@medsci.uu.se} \cr
Hanna Goransson Kultima, \email{hanna.goransson.kultima@medsci.uu.se} \cr
Sebastian DiLorenzo, \email{sebastian.dilorenzo@medsci.uu.se}
}
