\name{TAPS_call}
\alias{TAPS_call}
\alias{TAPS.call}
\alias{Taps.call}
\alias{Taps_call}


\title{
Tumor Abberation Prediction Suite
}

\description{
TAPS_call calls copynumbers and displays total and minor copynumbers of a tumor sample.
}

\usage{
TAPS_call(directory=NULL,xlim=c(-1,1),ylim=c(0,1),minseg=1,maxCn=12)
}

\arguments{
	\item{directory}{
	Default is getwd().
	}
	\item{xlim}{
	Default is c(-1,2).
	}
	\item{ylim}{
	Default is c(0,1).
	}
	\item{minseg}{
	Default is 1.
	}
	\item{maxCn}{
	Default is 12.
	}
}

\details{
TAPS_call is the second step in TAPS analysis. It takes the interpretation of plots generated by TAPS_plot as input.

%ADD MORE DETAILS HERE LATER!
}

\references{
http://www.biomedcentral.com/content/pdf/gb-2011-12-10-r108.pdf
}

\author{
Markus Mayrhofer, \email{markus.mayrhofer@medsci.uu.se} \cr
Hanna Goransson Kultima, \email{hanna.goransson.kultima@medsci.uu.se} \cr
Sebastian DiLorenzo, \email{sebastian.dilorenzo@medsci.uu.se}
}
