\name{TAPS_call}
\alias{TAPS_call}
\alias{TAPS.call}
\alias{Taps.call}
\alias{Taps_call}


\title{
Tumor Abberation Prediction Suite
}

\description{
TAPS_call calculates genome-wide total and minor copynumbers of tumor samples.
}

\usage{
TAPS_call(samples='all',directory=getwd(),cores=1)
}

\arguments{
	\item{samples}{
	Default is 'all'. Alternatively, supply a (boolean, integer or sample name) vector of included samples.
	}
	\item{directory}{
	Directory where the sample folders reside. Default is current directory.
	}
	\item{cores}{
	Default is 1. Set amount of threads/cores to be used.
	}
}

\details{
TAPS_call is the second step in the TAPS analysis, to be run after TAPS_plot. It requires: \cr
1) Your interpretation of plots which you add to the SampleData.csv file: \cr
    cn1-cn3: The log-ratio of copy number 1-3 (in the main clone), any two or three of these values are required. \cr
    loh: The allelic imbalance of copy number 2 LOH. \cr
    (Default values are fine for fairly pure, near-diploid samples) \cr
    See website for further instructions: http://patchwork.r-forge.r-project.org/ \cr
2) The .Rdata files previously generated by TAPS_plot().

%ADD MORE DETAILS HERE LATER!
}

\references{
http://www.biomedcentral.com/content/pdf/gb-2011-12-10-r108.pdf
}

\author{
Markus Mayrhofer, \email{markus.mayrhofer@medsci.uu.se} \cr
Hanna Goransson Kultima, \email{hanna.goransson.kultima@medsci.uu.se} \cr
Sebastian DiLorenzo, \email{sebastian.dilorenzo@medsci.uu.se} \cr
Bjorn Viklund, \email{bjorn.viklund@medsci.uu.se}
}
